+++ reapy/additional_api.py
@@ -3,9 +3,9 @@ from reapy import reascript_api as RPR
 from reapy.reascript_api import _RPR
 
 
-def packs_l(v: str) -> ct.c_char_p:
+def packs_l(v: str, encoding="latin-1") -> ct.c_char_p:
     MAX_STRBUF = 4 * 1024 * 1024
-    return ct.create_string_buffer(str(v).encode("latin-1"), MAX_STRBUF)
+    return ct.create_string_buffer(str(v).encode(encoding), MAX_STRBUF)
 
 
 def unpacks_l(v):
@@ -158,14 +158,16 @@ def MIDI_GetEvt(take, evtidx, selectedOut, mutedOut, ppqposOut, msg, msg_sz):
 #     r = f(t[0], byref(t[1]), byref(t[2]))
 #     return (r, p0, float(t[1].value), float(t[2].value))
 
-# def RPR_MIDI_GetHash(p0, p1, p2, p3):
-#     a = _ft['MIDI_GetHash']
-#     f = CFUNCTYPE(c_byte, c_uint64, c_byte, c_char_p, c_int)(a)
-#     t = (
-#         rpr_packp('MediaItem_Take*', p0), c_byte(p1), rpr_packs(p2), c_int(p3)
-#     )
-#     r = f(t[0], t[1], t[2], t[3])
-#     return (r, p0, p1, rpr_unpacks(t[2]), p3)
+def MIDI_GetHash(p0, p1, p2, p3):
+    a = _RPR._ft['MIDI_GetHash']
+    f = ct.CFUNCTYPE(ct.c_byte, ct.c_uint64, ct.c_byte,
+                     ct.c_char_p, ct.c_int)(a)
+    t = (
+        _RPR.rpr_packp('MediaItem_Take*',
+                       p0), ct.c_byte(p1), packs_l(p2), ct.c_int(p3)
+    )
+    r = f(t[0], t[1], t[2], t[3])
+    return (r, p0, p1, unpacks_l(t[2]), p3)
 
 # def RPR_MIDI_GetNote(p0, p1, p2, p3, p4, p5, p6, p7, p8):
 #     a = _ft['MIDI_GetNote']
@@ -257,12 +259,15 @@ def MIDI_GetEvt(take, evtidx, selectedOut, mutedOut, ppqposOut, msg, msg_sz):
 #         int(t[5].value), rpr_unpacks(t[6]), int(t[7].value)
 #     )
 
-# def RPR_MIDI_GetTrackHash(p0, p1, p2, p3):
-#     a = _ft['MIDI_GetTrackHash']
-#     f = CFUNCTYPE(c_byte, c_uint64, c_byte, c_char_p, c_int)(a)
-#     t = (rpr_packp('MediaTrack*', p0), c_byte(p1), rpr_packs(p2), c_int(p3))
-#     r = f(t[0], t[1], t[2], t[3])
-#     return (r, p0, p1, rpr_unpacks(t[2]), p3)
+
+def MIDI_GetTrackHash(p0, p1, p2, p3):
+    a = _RPR._ft['MIDI_GetTrackHash']
+    f = ct.CFUNCTYPE(ct.c_byte, ct.c_uint64, ct.c_byte,
+                     ct.c_char_p, ct.c_int)(a)
+    t = (_RPR.rpr_packp('MediaTrack*', p0),
+         ct.c_byte(p1), packs_l(p2), ct.c_int(p3))
+    r = f(t[0], t[1], t[2], t[3])
+    return (r, p0, p1, unpacks_l(t[2]), p3)
 
 # def RPR_MIDI_InsertCC(p0, p1, p2, p3, p4, p5, p6, p7):
 #     a = _ft['MIDI_InsertCC']
@@ -276,16 +281,23 @@ def MIDI_GetEvt(take, evtidx, selectedOut, mutedOut, ppqposOut, msg, msg_sz):
 #     r = f(t[0], t[1], t[2], t[3], t[4], t[5], t[6], t[7])
 #     return r
 
-# def RPR_MIDI_InsertEvt(p0, p1, p2, p3, p4, p5):
-#     a = _ft['MIDI_InsertEvt']
-#     f = CFUNCTYPE(c_byte, c_uint64, c_byte, c_byte, c_double, c_char_p,
-#                   c_int)(a)
-#     t = (
-#         rpr_packp('MediaItem_Take*', p0), c_byte(p1), c_byte(p2), c_double(p3),
-#         rpr_packsc(p4), c_int(p5)
-#     )
-#     r = f(t[0], t[1], t[2], t[3], t[4], t[5])
-#     return r
+
+def MIDI_InsertEvt(take, selected, muted, ppqpos, bytestr, bytestr_sz):
+    a = _RPR._ft['MIDI_InsertEvt']
+    f = ct.CFUNCTYPE(ct.c_byte, ct.c_uint64, ct.c_byte,
+                     ct.c_byte, ct.c_double, ct.c_char_p,
+                     ct.c_int)(a)
+    t = (
+        _RPR.rpr_packp('MediaItem_Take*',   # 0
+                       take),
+        ct.c_byte(selected),                # 1
+        ct.c_byte(muted),                   # 2
+        ct.c_double(ppqpos),                # 3
+        packs_l(bytestr),                   # 4
+        ct.c_int(bytestr_sz)                # 5
+    )
+    r = f(t[0], t[1], t[2], t[3], t[4], t[5])
+    return r
 
 # def RPR_MIDI_InsertNote(p0, p1, p2, p3, p4, p5, p6, p7, p8):
 #     a = _ft['MIDI_InsertNote']
@@ -300,17 +312,26 @@ def MIDI_GetEvt(take, evtidx, selectedOut, mutedOut, ppqposOut, msg, msg_sz):
 #     r = f(t[0], t[1], t[2], t[3], t[4], t[5], t[6], t[7], byref(t[8]))
 #     return (r, p0, p1, p2, p3, p4, p5, p6, p7, int(t[8].value))
 
-# def RPR_MIDI_InsertTextSysexEvt(p0, p1, p2, p3, p4, p5, p6):
-#     a = _ft['MIDI_InsertTextSysexEvt']
-#     f = CFUNCTYPE(
-#         c_byte, c_uint64, c_byte, c_byte, c_double, c_int, c_char_p, c_int
-#     )(a)
-#     t = (
-#         rpr_packp('MediaItem_Take*', p0), c_byte(p1), c_byte(p2), c_double(p3),
-#         c_int(p4), rpr_packsc(p5), c_int(p6)
-#     )
-#     r = f(t[0], t[1], t[2], t[3], t[4], t[5], t[6])
-#     return r
+
+def MIDI_InsertTextSysexEvt(take, selected, muted, ppqpos,
+                            type_, bytestr, bytestr_sz):
+    a = _RPR._ft['MIDI_InsertTextSysexEvt']
+    f = ct.CFUNCTYPE(
+        ct.c_byte, ct.c_uint64, ct.c_byte, ct.c_byte, ct.c_double,
+        ct.c_int, ct.c_char_p, ct.c_int
+    )(a)
+    t = (
+        _RPR.rpr_packp('MediaItem_Take*',   # 0
+                       take),
+        ct.c_byte(selected),                # 1
+        ct.c_byte(muted),                   # 2
+        ct.c_double(ppqpos),                # 3
+        ct.c_int(type_),                    # 4
+        packs_l(bytestr),                   # 5
+        ct.c_int(bytestr_sz)                # 6
+    )
+    r = f(t[0], t[1], t[2], t[3], t[4], t[5], t[6])
+    return r
 
 # def RPR_midi_reinit():
 #     a = _ft['midi_reinit']
@@ -360,24 +381,27 @@ def MIDI_GetEvt(take, evtidx, selectedOut, mutedOut, ppqposOut, msg, msg_sz):
 #     r = f(t[0], t[1], t[2], t[3], byref(t[4]))
 #     return (r, p0, p1, p2, p3, int(t[4].value))
 
-# def RPR_MIDI_SetEvt(p0, p1, p2, p3, p4, p5, p6, p7):
-#     a = _ft['MIDI_SetEvt']
-#     f = CFUNCTYPE(
-#         c_byte, c_uint64, c_int, c_void_p, c_void_p, c_void_p, c_char_p, c_int,
-#         c_void_p
-#     )(a)
-#     t = (
-#         rpr_packp('MediaItem_Take*', p0), c_int(p1), c_byte(p2), c_byte(p3),
-#         c_double(p4), rpr_packsc(p5), c_int(p6), c_byte(p7)
-#     )
-#     r = f(
-#         t[0], t[1], byref(t[2]), byref(t[3]), byref(t[4]), t[5], t[6],
-#         byref(t[7])
-#     )
-#     return (
-#         r, p0, p1, int(t[2].value), int(t[3].value), float(t[4].value), p5, p6,
-#         int(t[7].value)
-#     )
+
+def MIDI_SetEvt(p0, p1, p2, p3, p4, p5, p6, p7):
+    a = _RPR._ft['MIDI_SetEvt']
+    f = ct.CFUNCTYPE(
+        ct.c_byte, ct.c_uint64, ct.c_int, ct.c_void_p, ct.c_void_p,
+        ct.c_void_p, ct.c_char_p, ct.c_int,
+        ct.c_void_p
+    )(a)
+    t = (
+        _RPR.rpr_packp('MediaItem_Take*',
+                       p0), ct.c_int(p1), ct.c_byte(p2), ct.c_byte(p3),
+        ct.c_double(p4), packs_l(p5), ct.c_int(p6), ct.c_byte(p7)
+    )
+    r = f(
+        t[0], t[1], ct.byref(t[2]), ct.byref(t[3]), ct.byref(t[4]), t[5], t[6],
+        ct.byref(t[7])
+    )
+    return (
+        r, p0, p1, int(t[2].value), int(t[3].value), float(t[4].value), p5, p6,
+        int(t[7].value)
+    )
 
 # def RPR_MIDI_SetItemExtents(p0, p1, p2):
 #     a = _ft['MIDI_SetItemExtents']