+++ reapy/core/item/take.py
@@ -30,6 +30,32 @@ class Take(ReapyObject):
         audio_accessor = reapy.AudioAccessor(audio_accessor_id)
         return audio_accessor
 
+    @reapy.inside_reaper()
+    def add_event(self, message, position, unit="seconds"):
+        """
+        Add generic event to the take at position.
+
+        Note
+        ----
+        ⋅ No sort events during this call
+        ⋅ For a long time inserting a notes within this function
+        caused problems. Better to use the proper function for that.
+
+        Parameters
+        ----------
+        message : Iterable[int]
+            Can be any message buffer, for example: (0xb0, 64, 127)
+            which is CC64 val127 on channel 1
+        position : float
+            position at take
+        unit : str, optional
+            "beats"|"ppq"|"seconds" (default are seconds)
+        """
+        ppqpos = self._resolve_midi_unit((position,), unit)[0]
+        bytestr = self._midi_to_bytestr(message)
+        RPR.MIDI_InsertEvt(self.id, False, False,
+                           ppqpos, bytestr, len(bytestr))
+
     def add_fx(self, name, even_if_exists=True):
         """
         Add FX to track and return it.
@@ -108,6 +134,36 @@ class Take(ReapyObject):
         )
         RPR.MIDI_InsertNote(*args)
 
+    @reapy.inside_reaper()
+    def add_sysex(self, message, position, unit="seconds", evt_type=-1):
+        """
+        Add SysEx event to take.
+
+        Note
+        ----
+        ⋅ No sort events during this call
+        ⋅ No need to add 0xf0 ... 0xf7 bytes (they will be doubled)
+
+        Parameters
+        ----------
+        message : Iterable[int]
+            Can be any message buffer, for example: (0xb0, 64, 127)
+            which is CC64 val127 on channel 1
+        position : float
+            position at take
+        unit : str, optional
+            "beats"|"ppq"|"seconds" (default are seconds)
+        evt_type: int (default -1)
+            Allowable types are
+            ⋅ -1:sysex (msg should not include bounding F0..F7),
+            ⋅ 1-14:MIDI text event types,
+            ⋅ 15=REAPER notation event.
+        """
+        bytestr = self._midi_to_bytestr(message)
+        ppqpos = self._resolve_midi_unit((position,), unit)[0]
+        RPR.MIDI_InsertTextSysexEvt(self.id, False, False, ppqpos,
+                                    evt_type, bytestr, len(bytestr))
+
     def beat_to_ppq(self, beat):
         """
         Convert beat number (from project start) to MIDI ticks (of the take).
@@ -185,7 +241,8 @@ class Take(ReapyObject):
 
     @property
     def guid(self):
-        """Used for communication within other scripts.
+        """
+        Used for communication within other scripts.
 
         :type: str
         """
@@ -198,6 +255,37 @@ class Take(ReapyObject):
         Make take active.
         """
         RPR.SetActiveTake(self.id)
+
+    @property
+    def midi_events(self):
+        """
+        Get all midi events as EventList.
+
+        Returns
+        -------
+        MIDIEventList
+        """
+        return reapy.core.item.midi_event.MIDIEventList(self)
+
+    def midi_hash(self, notes_only=False):
+        """
+        Get hash of MIDI-data to compare with later.
+
+        Parameters
+        ----------
+        notes_only : bool, (False by default)
+            count only notes if True
+
+        Returns
+        -------
+        str
+        """
+        _, _, _, hash_, _ = RPR.MIDI_GetHash(
+            self.id, notes_only, 'hash', 1024*1024)
+        return hash_
+
+    def _midi_to_bytestr(self, message):
+        return bytes(message).decode('latin-1')
 
     @property
     def n_cc(self):
@@ -225,6 +313,15 @@ class Take(ReapyObject):
         :type: int
         """
         return RPR.TakeFX_GetCount(self.id)
+
+    @property
+    def n_midi_events(self):
+        """
+        Number of MIDI events in take.
+
+        :type: int
+        """
+        return RPR.MIDI_CountEvts(self.id, 1, 1, 1)[0]
 
     @property
     def n_notes(self):